Animation Fixes
Working:

- Tower

-- Damage Animation

- Blob

-- Walking

-- Attacking

-- Dieing

Not Working:

- Tower

-- Door Animation

Tower Damage animation now gets it's component via transform (not gameobject), to ensure it only get's it's own components.

Tower Door animation is seemingly untriggerable and if you trigger it by hand, the door suddenly moves from the side to the front of the tower, leaving a hole in the side.

Blobs now use the walking animation. Except for the green blob who jitters a bit upon reaching his destination, movement is smooth.

Blobs now use the attack animation when close in range.

Blobs now explode into a bunch of small blobs, when dieing. Dieing animation is now triggered by the currenthealth of the corresponding

HealthObject of the Blob and not by the 'die'-trigger anymore.

The animation has 2 seconds before the object is destroyed.

Most AnimationScripts are now NetworkBehaviours and have SyncVars, enabling the animation across network

Changed the Selector debug boolean VARIABLE to an actual function call => Debug.isDebugBuild (not tested if that works)

Changed the code flow when checking for enemies. Break ensures the Blob only ever attacks one target at a time.

Commented out a couple of log outputs for easier readibility for me and removed a bit of unused code.